requirements.txt
click==8.0.1
Flask==2.0.1
Flask-WTF==0.15.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
numpy==1.21.2
py-bcrypt==0.4
scikit-learn==0.24.2
scipy==1.7.1
SharedArray==3.2.1
threadpoolctl==2.2.0
Werkzeug==2.0.1
WTForms==2.3.3
